changed structure and moved di to its respective presentation modules
