+++ app/src/main/java/com/andonichc/postsapp/PostApplication.kt
@@ -3,12 +3,10 @@ package com.andonichc.postsapp
 import android.app.Application
 import com.andonichc.postsapp.di.component.DaggerAppComponent
 import com.andonichc.postsapp.di.component.DaggerNetComponent
-import com.andonichc.postsapp.di.module.AppModule
-import com.andonichc.postsapp.di.module.NetModule
+import com.andonichc.postsapp.presentation.base.di.AppModule
+import com.andonichc.postsapp.presentation.base.di.NetModule
 
-/**
- * Created by andoni on 19/12/2017.
- */
+
 class PostApplication : Application() {
     val mAppComponent = DaggerAppComponent.builder()
             .appModule(AppModule(this))
