+++ app/src/main/java/com/andonichc/postsapp/presentation/base/di/AppModule.kt
@@ -1,4 +1,4 @@
-package com.andonichc.postsapp.di.module
+package com.andonichc.postsapp.presentation.base.di
 
 import com.andonichc.postsapp.PostApplication
 import dagger.Module
